input variable
